// src/app/models/master.model.ts
import {BaseTableModel} from './base-table.model';

export class MasterModel extends BaseTableModel {
    MarkInit: string;
    Mark: string;
    DeliveringRoad: string;
    ReceivingRoad: string;
    Splc: string;
    InterchangeType: string;
    City: string;
    CityName: string;
    State: string;
    StateName: string;
    ReasonCode: string;
    ChangeReason: string;
}